Fix to Variable.lessThen when both are unbounded variables

+++ src/com/ugos/jiprolog/engine/PrettyPrinter.java
@@ -561,7 +561,7 @@ final class PrettyPrinter extends Object
             return var.getName();
         }
         else if(object == null)
-            return "_" + Integer.toString(var.getAddress());// + ":" + Integer.toString(var.hashCode()); /* + var.getName() + "?";*/
+            return "_" + Long.toString(var.getAddress());// + ":" + Integer.toString(var.hashCode()); /* + var.getName() + "?";*/
             //return var.getName()  + " = " + "_" + Integer.toString(var.getAddress());// + ":" + Integer.toString(var.hashCode()); /* + var.getName() + "?";*/
         else
             varTable.put(var.getName(), var);
